+++ src/lib/global.js
@@ -4,11 +4,12 @@ export function getGlobal() {
   return {
     debug: false,
     transitions: true,
+    actions: {},
+    errors: {},
     parse: function () {
       // Parse the DOM for zjax elements
       parseSwaps(document);
       parseActions(document);
     },
-    actions: {},
   };
 }