+++ src/ui.js
@@ -1,14 +1,12 @@
 document.addEventListener("DOMContentLoaded", () => {
   zjax.debug = true;
 
-  zjax.actions = {
-    ui: {
-      alert() {
-        alert("You clicked the b key.");
-      },
-      gotoFacebook($) {
-        $.redirect("https://facebook.com");
-      },
+  zjax.actions.ui = {
+    alert() {
+      alert("You clicked the b key.");
+    },
+    gotoFacebook($) {
+      $.redirect("https://facebook.com");
     },
   };
 });